# ARROW-11268: [Rust][DataFusion] MemTable::load output partition support