--- a/rust/datafusion/benches/sort_limit_query_sql.rs
+++ b/rust/datafusion/benches/sort_limit_query_sql.rs
@@ -70,8 +70,11 @@ fn create_context() -> Arc<Mutex<ExecutionContext>> {
 
     let ctx_holder: Arc<Mutex<Vec<Arc<Mutex<ExecutionContext>>>>> =
         Arc::new(Mutex::new(vec![]));
+
+    let partitions = 16;
+
     rt.block_on(async {
-        let mem_table = MemTable::load(&csv, 16 * 1024).await.unwrap();
+        let mem_table = MemTable::load(&csv, 16 * 1024, Some(partitions)).await.unwrap();
 
         // create local execution context
         let mut ctx = ExecutionContext::new();