@Dandandan 2021-01-17
This would also help us in the db-benchmark https://github.com/h2oai/db-benchmark/pull/182